refactor: constants 파일을 errorMessages로 변경
